projects/streaming_app/pyproject.toml
[project]
name = "streaming-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "kafka-python (>=2.3.0,<3.0.0)",
    "confluent-kafka (>=2.13.0,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pyspark (>=4.1.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
